app/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
PyPDF2
pytesseract
pdf2image
google-generativeai
Pillow 